% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Eur2014Dpts}
\alias{Eur2014Dpts}
\title{Résultats des élections européennes de 2014, par département}
\format{un jeu de données de 107 lignes et 234 variables}
\source{
\url{http://www.data.gouv.fr/fr/dataset/elections-europeennes-2014-resulta-1}
}
\description{
Résultats des élections européennes de 2014, par département
}
\details{
Un jeu de données contenant les résultats des élections européennes de 2014, par département. Les données ont été téléchargées sur \url{http://www.data.gouv.fr/fr/dataset/elections-europeennes-2014-resulta-1}, puis la feuille consacrée aux départements a été sauvegardée en csv (en enlevant simplement les deux premières lignes, pour que les titres de colonnes soient sur la première ligne). Elle a ensuite été importée dans R avec le code suivant : \code{Eur2014Dpts <- read.csv("./Eur2014Dpts.csv", header=TRUE, sep=";", stringsAsFactor=FALSE)}.
}
\keyword{dataset}

