% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lire}
\alias{lire}
\title{Fonction pour transformer les fichiers du ministère de l'intérieur en data.frame utilisable}
\usage{
lire(X, keep, col, keep.names = names(res), gap = 3)
}
\arguments{
\item{X}{un data.frame brut. Généralement importé depuis un CSV.}

\item{keep}{vecteur indiquant le numéro (ou le nom) des colonnes à conserver telles quelles (habituellement Inscrits, Exprimés, Blancs, Nuls, etc.). Il doit impérativement y avoir les Inscrits et les Exprimés.}

\item{col}{vecteur indiquant le numéro des colonnes contenant les étiquettes à partir desquelles les résultats sont agrégés. Généralement, leur espacement est régulier, donc peut s'écrire sous la forme de, e.g., \code{seq(10, 100, 10)}.}

\item{keep.names}{le nom à donner aux colonnes conservées en l'état (paramètre \code{keep}). Par défaut, le nom de ces colonnes dans le \code{data.frame X}. Il est nécessaire qu'il y ait une colonne intitulée "Inscrits" et une colonne intitulée "Exprimés".}

\item{gap}{décalage entre les colonnes avec les étiquettes et les colonnes avec le nombre de voix. Par défaut, 3.}
}
\value{
un \code{data.frame} avec les colonnes conservées ainsi que, pour chaque nuance politique, une colonne avec le total des voix, une colonne avec le score rapporté aux inscrits, une colonne avec le score rapporté aux exprimés. Enfin, une colonne indique le nombre de candidats (ou de listes) qu'il y avait dans la circonscription territoriale.
}
\description{
La fonction \code{lire} est utile pour transformer les fichiers de résultats électoraux diffusés par le ministère de l'Intérieur français, lorsque l'offre électorale n'est pas homogène sur l'ensemble du territoire (législatives, européennes, cantonales, régionales...). Pour ce faire, les résultats sont agrégés en fonction des étiquettes attribuées par le ministère de l'intérieur.
}
\examples{
data(Eur2014Dpts)
 names(Eur2014Dpts)
 res <- lire(Eur2014Dpts, keep = c(2,4,5,7,9,12,15), col=c(seq(19,229,7)), keep.names=c("CodeDpt", "Inscrits", "Asbtentions","Votants", "Blancs", "Nuls", "Exprimés"))
}

