% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LireMinInterieur.R
\name{lireInteractif}
\alias{lireInteractif}
\title{Version interactive de la fonction \code{lire}}
\usage{
lireInteractif()
}
\description{
Cette fonction lance une application shiny qui permet d'utiliser de manière interactive la fonction \code{lire}.
}

