% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{LireMinInterieur}
\alias{LireMinInterieur}
\alias{LireMinInterieur-package}
\title{Ce package offre des outils pour nettoyer les fichiers électoraux du ministère de l'intérieur.}
\description{
Ce package offre des outils pour nettoyer les fichiers électoraux du ministère de l'intérieur.
}
\author{
Joel Gombin \email{joel.gombin@gmail.com}
}
\keyword{package}

